Make histogram usable again
